backend/requirements.txt
# GDNA Lyzr Baseline - Essential Backend Dependencies
# Clean, focused backend for Lyzr agent deployment

# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0

# System monitoring
psutil>=5.9.0

# Database drivers (for health checks only)
psycopg2-binary>=2.9.0
pymongo>=4.6.0
redis>=5.0.0
pika>=1.3.0
opensearch-py>=2.0.0

# HTTP client for health checks
aiohttp>=3.9.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0